refactor: reorganize Room module, remove SQLDelight, fix review issues
- Reorganize Room files into subpackages (daos/, entities/, converters/, di/, paging/)
- Remove SQLDelight module entirely
- Fix iOS provider crash when directory already exists
- Fix unsafe cast in MappingPagingSource.getRefreshKey
- Make upsert methods return Long/List<Long> instead of Unit
- Use conditional insert in upsert (skip insert when update succeeds)
- Add search history upsert timestamp update test
- Fix detekt violations (SnackbarMessage lambda capture, AppUi name shadowing)
