+++ data/database-room/src/commonMain/kotlin/dev/avatsav/linkding/data/db/daos/RoomPagingBookmarksDao.kt
@@ -2,9 +2,9 @@ package dev.avatsav.linkding.data.db.daos
 
 import androidx.paging.PagingSource
 import dev.avatsav.linkding.data.db.room.LinkdingRoomDatabase
-import dev.avatsav.linkding.data.db.room.MappingPagingSource
-import dev.avatsav.linkding.data.db.room.toModel
-import dev.avatsav.linkding.data.db.room.toRoomEntity
+import dev.avatsav.linkding.data.db.room.entities.toModel
+import dev.avatsav.linkding.data.db.room.entities.toRoomEntity
+import dev.avatsav.linkding.data.db.room.paging.MappingPagingSource
 import dev.avatsav.linkding.data.model.Bookmark
 import dev.zacsweers.metro.AppScope
 import dev.zacsweers.metro.ContributesBinding
@@ -18,9 +18,8 @@ class RoomPagingBookmarksDao(private val db: LinkdingRoomDatabase) : PagingBookm
   override fun bookmarksPagingSource(): PagingSource<Int, Bookmark> =
     MappingPagingSource(db.bookmarksQueries().bookmarksPagingSource(), mapper = { it.toModel() })
 
-  override suspend fun refresh(bookmarks: List<Bookmark>) {
+  override suspend fun refresh(bookmarks: List<Bookmark>) =
     db.bookmarksQueries().replaceAll(bookmarks.map { it.toRoomEntity() })
-  }
 
   override suspend fun append(bookmarks: List<Bookmark>) {
     db.bookmarksQueries().upsertAll(bookmarks.map { it.toRoomEntity() })
