+++ data/database-room/src/commonMain/kotlin/dev/avatsav/linkding/data/db/daos/RoomSearchHistoryDao.kt
@@ -1,8 +1,8 @@
 package dev.avatsav.linkding.data.db.daos
 
 import dev.avatsav.linkding.data.db.room.LinkdingRoomDatabase
-import dev.avatsav.linkding.data.db.room.toModel
-import dev.avatsav.linkding.data.db.room.toRoomEntity
+import dev.avatsav.linkding.data.db.room.entities.toModel
+import dev.avatsav.linkding.data.db.room.entities.toRoomEntity
 import dev.avatsav.linkding.data.model.SearchHistory
 import dev.zacsweers.metro.AppScope
 import dev.zacsweers.metro.ContributesBinding
@@ -15,22 +15,18 @@ import kotlinx.coroutines.flow.map
 @SingleIn(AppScope::class)
 @ContributesBinding(AppScope::class)
 class RoomSearchHistoryDao(private val db: LinkdingRoomDatabase) : SearchHistoryDao {
-  override suspend fun upsert(history: SearchHistory) {
+  override suspend fun upsert(history: SearchHistory): Long =
     db.searchHistoryQueries().upsert(history.toRoomEntity())
-  }
 
-  override suspend fun insertAll(histories: List<SearchHistory>) {
+  override suspend fun insertAll(histories: List<SearchHistory>) =
     db.searchHistoryQueries().insertAll(histories.map { it.toRoomEntity() })
-  }
 
   override fun observeRecent(limit: Long): Flow<List<SearchHistory>> =
-    db.searchHistoryQueries().observeRecent(limit).map { histories -> histories.map { it.toModel() } }
+    db.searchHistoryQueries().observeRecent(limit).map { histories ->
+      histories.map { it.toModel() }
+    }
 
-  override suspend fun deleteAll() {
-    db.searchHistoryQueries().deleteAll()
-  }
+  override suspend fun deleteAll() = db.searchHistoryQueries().deleteAll()
 
-  override suspend fun pruneToLimit(limit: Long) {
-    db.searchHistoryQueries().pruneToLimit(limit)
-  }
+  override suspend fun pruneToLimit(limit: Long) = db.searchHistoryQueries().pruneToLimit(limit)
 }
