--- data/database-room/src/commonMain/kotlin/dev/avatsav/linkding/data/db/room/RoomProviders.kt
@@ -1,3 +1,0 @@
-package dev.avatsav.linkding.data.db.room
-
-expect interface RoomPlatformProviders
