+++ data/database-room/src/commonMain/kotlin/dev/avatsav/linkding/data/db/room/di/RoomProviders.kt
@@ -1,0 +1,3 @@
+package dev.avatsav.linkding.data.db.room.di
+
+expect interface RoomPlatformProviders
