--- data/database-sqldelight/build.gradle.kts
@@ -1,27 +1,0 @@
-plugins {
-  alias(libs.plugins.sqldelight)
-  id("convention.kmp.lib")
-}
-
-kotlin {
-  sourceSets {
-    commonMain.dependencies {
-      implementation(projects.core.base)
-      api(projects.data.database) // Implements database contracts
-      api(libs.sqldelight.coroutines)
-      implementation(libs.sqldelight.primitive)
-      implementation(projects.data.models) // Internal use, models exposed via database contract
-      implementation(libs.kotlin.datetime)
-    }
-    androidMain.dependencies { implementation(libs.sqldelight.android) }
-    iosMain.dependencies {
-      // Required for iOS build:
-      // https://github.com/cashapp/sqldelight/issues/4888#issuecomment-1846036472
-      implementation(libs.stately)
-      implementation(libs.sqldelight.native)
-    }
-    jvmMain.dependencies { implementation(libs.sqldelight.sqlite) }
-  }
-}
-
-sqldelight { databases { create("Database") { packageName = "dev.avatsav.linkding.data.db" } } }
