--- data/database-sqldelight/src/androidMain/kotlin/dev/avatsav/linkding/data/db/di/SqlDelightPlatformProviders.kt
@@ -1,17 +1,0 @@
-package dev.avatsav.linkding.data.db.di
-
-import android.app.Application
-import app.cash.sqldelight.db.SqlDriver
-import app.cash.sqldelight.driver.android.AndroidSqliteDriver
-import dev.avatsav.linkding.data.db.Database
-import dev.zacsweers.metro.AppScope
-import dev.zacsweers.metro.Provides
-import dev.zacsweers.metro.SingleIn
-
-actual interface SqlDelightPlatformProviders {
-
-  @Provides
-  @SingleIn(AppScope::class)
-  fun provideSqliteDriver(application: Application): SqlDriver =
-    AndroidSqliteDriver(schema = Database.Schema, context = application, name = "linkding.db")
-}
