--- data/database-sqldelight/src/commonMain/kotlin/dev/avatsav/linkding/data/db/adapters/InstantColumnAdapter.kt
@@ -1,10 +1,0 @@
-package dev.avatsav.linkding.data.db.adapters
-
-import app.cash.sqldelight.ColumnAdapter
-import kotlin.time.Instant
-
-internal object InstantColumnAdapter : ColumnAdapter<Instant, String> {
-  override fun encode(value: Instant): String = value.toString()
-
-  override fun decode(databaseValue: String): Instant = Instant.parse(databaseValue)
-}
