--- data/database-sqldelight/src/commonMain/sqldelight/dev/avatsav/linkding/data/db/search_history.sq
@@ -1,46 +1,0 @@
-import kotlin.time.Instant;
-
-CREATE TABLE IF NOT EXISTS `search_history` (
-    `id` INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL,
-    `query` TEXT NOT NULL,
-    `modified` TEXT AS Instant NOT NULL
-);
-
-CREATE UNIQUE INDEX IF NOT EXISTS `index_search_history_unique` ON `search_history` (`query`);
-
-selectRecent:
-SELECT *
-FROM search_history
-ORDER BY `modified` DESC
-LIMIT :limit;
-
-insert:
-INSERT INTO search_history (
-    `query`,
-    `modified`
-) VALUES (
-    :query,
-    :modified
-);
-
-updatemodified:
-UPDATE search_history SET `modified` = :modified
-WHERE `query` = :query;
-
-upsert {
-  UPDATE search_history SET `modified` = :modified
-  WHERE `query` = :query;
-  INSERT OR IGNORE INTO search_history (
-    `query`, `modified`
-  ) VALUES (
-    :query, :modified
-  );
-}
-
-deleteAll:
-DELETE FROM search_history;
-
-pruneBeyond:
-DELETE FROM search_history WHERE `id` NOT IN (
-  SELECT `id` FROM search_history ORDER BY `modified` DESC LIMIT :limit
-);
