--- data/database-sqldelight/src/iosMain/kotlin/dev/avatsav/linkding/data/db/di/SqlDelightPlatformProviders.kt
@@ -1,16 +1,0 @@
-package dev.avatsav.linkding.data.db.di
-
-import app.cash.sqldelight.db.SqlDriver
-import app.cash.sqldelight.driver.native.NativeSqliteDriver
-import dev.avatsav.linkding.data.db.Database
-import dev.zacsweers.metro.AppScope
-import dev.zacsweers.metro.Provides
-import dev.zacsweers.metro.SingleIn
-
-actual interface SqlDelightPlatformProviders {
-
-  @Provides
-  @SingleIn(AppScope::class)
-  fun provideSqliteDriver(): SqlDriver =
-    NativeSqliteDriver(schema = Database.Schema, name = "linkding.db")
-}
